[DX] Extract refund unit total provider
