+++ src/Resources/config/services/providers.xml
@@ -38,9 +38,13 @@
             class="Sylius\RefundPlugin\Provider\CurrentDateTimeImmutableProvider"
         />
 
-        <service id="Sylius\RefundPlugin\Provider\RemainingTotalProviderInterface" class="Sylius\RefundPlugin\Provider\RemainingTotalProvider">
+        <service id="Sylius\RefundPlugin\Provider\RefundUnitTotalProviderInterface" class="Sylius\RefundPlugin\Provider\RefundUnitTotalProvider">
             <argument type="service" id="sylius.repository.order_item_unit" />
             <argument type="service" id="sylius.repository.adjustment" />
+        </service>
+
+        <service id="Sylius\RefundPlugin\Provider\RemainingTotalProviderInterface" class="Sylius\RefundPlugin\Provider\RemainingTotalProvider">
+            <argument type="service" id="Sylius\RefundPlugin\Provider\RefundUnitTotalProviderInterface" />
             <argument type="service" id="sylius_refund.repository.refund" />
         </service>
         <service id="Sylius\RefundPlugin\Provider\RemainingTotalProvider" alias="Sylius\RefundPlugin\Provider\RemainingTotalProviderInterface">
